+++ include/hyperion/ImageProcessor.h
@@ -141,13 +141,19 @@ public:
 				colors = _imageToLedColors->getUniLedColor(image);
 				break;
 			case 2:
-				colors = _imageToLedColors->getMeanLedColorSqrt(image);
+				colors = _imageToLedColors->getMeanSqrtLedColor(image);
 				break;
 			case 3:
 				colors = _imageToLedColors->getDominantLedColor(image);
 				break;
 			case 4:
-				colors = _imageToLedColors->getDominantLedColorAdv(image);
+				colors = _imageToLedColors->getDominantUniLedColor(image);
+				break;
+			case 5:
+				colors = _imageToLedColors->getDominantAdvLedColor(image);
+				break;
+			case 6:
+				colors = _imageToLedColors->getDominantAdvUniLedColor(image);
 				break;
 			default:
 				colors = _imageToLedColors->getMeanLedColor(image);
@@ -186,14 +192,21 @@ public:
 				_imageToLedColors->getUniLedColor(image, ledColors);
 				break;
 			case 2:
-				_imageToLedColors->getMeanLedColorSqrt(image, ledColors);
+				_imageToLedColors->getMeanSqrtLedColor(image, ledColors);
 				break;
 			case 3:
 				_imageToLedColors->getDominantLedColor(image, ledColors);
 				break;
 			case 4:
-				_imageToLedColors->getDominantLedColorAdv(image, ledColors);
+				_imageToLedColors->getDominantUniLedColor(image, ledColors);
 				break;
+			case 5:
+				_imageToLedColors->getDominantAdvLedColor(image, ledColors);
+				break;
+			case 6:
+				_imageToLedColors->getDominantAdvUniLedColor(image, ledColors);
+				break;
+
 			default:
 				_imageToLedColors->getMeanLedColor(image, ledColors);
 			}